[TM-920] Create a PointGeometry PHP model class and share the geojson attribute getters from PolygonGeometry.
